- bulk caption loading
